Make song counter more dynamic
Features:
Made the labels in the song counter box not be hard coded. Instead the values next to the labels are calculated at the start of each game
Added a song count

Bug Fixes:
Fixed a bug where song count reported the wrong number when entering in the middle of a game.